Vexing Vehicle Purchase problem.
In this exercise, you are tasked to write some code to help you prepare to buy a vehicle.

You have three tasks: one to determine if you will need one to help you choose between two vehicles, one to determine which license you will need, and one to estimate the acceptable price for a used vehicle.